Arreglando los puntos del examen

+++ Entidades/Aportes.cs
@@ -14,6 +14,6 @@ namespace P1_AP1_Vismar_20190425.Entidades
         public DateTime Fecha { get; set; } = DateTime.Now;
         public string Persona { get; set; }
         public string Concepto { get; set; }
-        public int Monto { get; set; }
+        public float Monto { get; set; }
     }
 }
